adding quotes on output path, in case of blank space breaking ffmpeg command
I had a runtime error of "unable to find a suitable output format for '{some parts of my data path}'" when running the example, and figured out the reason was that my openframeworks was placed inside a folder named with space character. It broke the output part of the FFmpeg command by chopping half of it and treat the other half as another flag. Putting a pair of quotes on the path saved my butt. 👍